+++ src/utilities/mod.rs
@@ -1,5 +1,1 @@
-//! This module contains utility functions to help you generate keys,
-//! as well as decode keys from different formats.
-
-pub(crate) mod base64;
-pub mod key_generation;
+pub mod base64;