+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 - Removed the `utilities::key_decoder` module, as it's out of scope for this library.
+- Removed the `utilities::key_generation` module, as it's out of scope for this library.
 - Removed the `with_new_payload` & `with_new_payload_and_key_config` methods as they add no value.
 
 ## [0.2.0] - 2024-04-23